Update the media content block to use the new media parallax component
