+++ src/blocks/CaseStudyParallax/index.module.scss
@@ -325,15 +325,3 @@
   }
 }
 
-.parallaxMedia {
-  position: relative;
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: 1fr;
-
-
-  & .parallaxItem  {
-    grid-row: 1/2;
-    grid-column: 1/2;
-  }
-}
